feat: create MovieCard and styles
